+++ lib/src/flutter/widget/src/text_form_field.dart
@@ -1,9 +1,8 @@
 import "package:flutter/material.dart";
-import "package:quickly/quickly.dart";
 
-/// A custom text form field widget with additional features and improved functionality.
+import "../../../../quickly.dart";
+
 class FxTextFormField extends StatefulWidget {
-  /// Constructs a FxTextFormField.
   const FxTextFormField({
     super.key,
     this.hintText,
@@ -29,67 +28,26 @@ class FxTextFormField extends StatefulWidget {
     this.textCapitalization = TextCapitalization.none,
   });
 
-  /// The label displayed above the text form field.
   final String? label;
-
-  /// The hint text displayed inside the text form field when empty.
   final String? hintText;
-
-  /// The background color of the text form field.
   final Color? fieldColor;
-
-  /// The border color of the text form field.
   final Color? borderColor;
-
-  /// The icon displayed as a prefix of the text form field.
   final IconData? prefixIcon;
-
-  /// The icon displayed as a suffix of the text form field.
   final IconData? suffixIcon;
-
-  /// Determines whether the text form field is secure (e.g., for passwords).
   final bool isSecure;
-
-  /// The border radius of the text form field container.
   final BorderRadiusGeometry? borderRadius;
-
-  /// The maximum number of lines for multi-line input.
   final int? maxLines;
-
-  /// The type of keyboard to display for text input.
   final TextInputType? keyboardType;
-
-  /// Determines whether to display a border around the text form field.
   final bool border;
-
-  /// The controller for editing the text.
   final TextEditingController? controller;
-
-  /// The list of validation rules to apply to the text form field.
   final List<String>? validations;
-
-  /// The padding around the text form field container.
   final EdgeInsets? padding;
-
-  /// The box shadow for the text form field container.
   final List<BoxShadow>? boxShadow;
-
-  /// Called when the text changes.
   final ValueChanged<String>? onChanged;
-
-  /// Defines the keyboard focus for this widget.
   final FocusNode? focusNode;
-
-  /// Whether the form field is enabled.
   final bool enabled;
-
-  /// The type of action button to use for the keyboard.
   final TextInputAction? textInputAction;
-
-  /// The style to use for the text being edited.
   final TextStyle? style;
-
-  /// Configures how the platform keyboard will select an uppercase or lowercase keyboard.
   final TextCapitalization textCapitalization;
 
   @override
@@ -127,18 +85,15 @@ class _FxTextFormFieldState extends State<FxTextFormField> {
             padding: const EdgeInsets.only(bottom: 8.0),
             child: Text(
               widget.label!,
-              style: const TextStyle(
-                  fontSize: 16,
-                  fontWeight: FontWeight.w500,
-                  color: FxColor.kcText),
+              style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                    fontWeight: FontWeight.w500,
+                    color: Colors.black,
+                  ),
             ),
           ),
         Container(
           decoration: BoxDecoration(
-            color: widget.fieldColor ?? FxColor.gray100,
-            border: widget.border
-                ? Border.all(color: widget.borderColor ?? FxColor.gray300)
-                : null,
+            color: widget.fieldColor ?? Colors.grey[100],
             borderRadius: widget.borderRadius ?? BorderRadius.circular(8),
             boxShadow: widget.boxShadow,
           ),
@@ -163,13 +118,16 @@ class _FxTextFormFieldState extends State<FxTextFormField> {
               prefixIcon:
                   widget.prefixIcon != null ? Icon(widget.prefixIcon) : null,
               hintText: widget.hintText,
+              hintStyle: Theme.of(context).textTheme.bodySmall?.copyWith(
+                    color: Colors.grey[600],
+                  ),
               border: InputBorder.none,
             ),
             keyboardType: widget.keyboardType,
             obscureText: _obscureText,
             maxLines: widget.isSecure ? 1 : widget.maxLines,
             enabled: widget.enabled,
-            style: widget.style,
+            style: widget.style ?? Theme.of(context).textTheme.bodyLarge,
             textInputAction: widget.textInputAction,
             textCapitalization: widget.textCapitalization,
             onChanged: (value) {
@@ -182,6 +140,7 @@ class _FxTextFormFieldState extends State<FxTextFormField> {
               }
               if (widget.validations != null) {
                 for (final validation in widget.validations!) {
+                  // Assuming you have a list of validation functions
                   final validate = validators[validation];
                   if (validate != null) {
                     final result = validate(value);